Make credential login a two-stage process
