+++ docs/contribute/workflows/README.md
@@ -2,8 +2,11 @@
 
 A collection of sequence diagrams for key parts of the FlowForge platform
 
+ - [Login](../workflows/login.md)
  - [User Sign Up](../workflows/signup.md)
  - [Password Reset](../workflows/password-reset.md)
- - [Invite ExternalUser](../workflows/invite-external-user.md)
+ - [Invite External User](../workflows/invite-external-user.md)
  - [Team Create](../workflows/team-create.md)
  - [Project Create](../workflows/project-create.md)
+ - [Project States](../workflows/project-states.md)
+
