+++ frontend/src/store/account.js
@@ -132,10 +132,6 @@ const mutations = {
     },
     setOffline (state, value) {
         state.offline = value
-    },
-    userSuspended (state, error) {
-        state.loginInflight = false
-        state.loginError = error
     }
 }
 
@@ -252,9 +248,9 @@ const actions = {
             state.dispatch('checkState', state.getters.redirectUrlAfterLogin)
         } catch (err) {
             if (err.response.status === 401) {
-                state.commit('loginFailed', 'Login failed')
+                state.commit('loginFailed', { code: 'login_failed', error: 'Login failed' })
             } else if (err.response.status === 403) {
-                state.commit('userSuspended', err.response.data.error)
+                state.commit('loginFailed', err.response.data)
             }
         }
     },
